+++ src/JestExt.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { ProjectWorkspace, JestTotalResults } from 'jest-editor-support';
 
-import * as decorations from './decorations';
+import { Decorations } from './Decorations';
 import { PluginResourceSettings } from './Settings';
 import { statusBar, Status, StatusBar, Mode } from './StatusBar';
 import {
@@ -50,17 +50,12 @@ export class JestExt {
 
   private jestWorkspace: ProjectWorkspace;
   private pluginSettings: PluginResourceSettings;
-  private context: vscode.ExtensionContext;
+  private decorations: Decorations;
   private workspaceFolder: vscode.WorkspaceFolder;
   private instanceSettings: InstanceSettings;
 
   // The ability to show fails in the problems section
   private failDiagnostics: vscode.DiagnosticCollection;
-
-  private passingItStyle: vscode.TextEditorDecorationType;
-  private failingItStyle: vscode.TextEditorDecorationType;
-  private skipItStyle: vscode.TextEditorDecorationType;
-  private unknownItStyle: vscode.TextEditorDecorationType;
 
   private parsingTestFile = false;
 
@@ -83,7 +78,6 @@ export class JestExt {
     instanceSettings: InstanceSettings,
     coverageCodeLensProvider: CoverageCodeLensProvider
   ) {
-    this.context = context;
     this.workspaceFolder = workspaceFolder;
     this.jestWorkspace = jestWorkspace;
     this.channel = outputChannel;
@@ -115,7 +109,7 @@ export class JestExt {
     this.handleJestEditorSupportEvent = this.handleJestEditorSupportEvent.bind(this);
 
     // The theme stuff
-    this.setupDecorators();
+    this.decorations = new Decorations(context);
     // The bottom bar thing
     this.setupStatusBar();
     // reset the jest diagnostics
@@ -225,26 +219,26 @@ export class JestExt {
   }
 
   updateDecorators(testResults: SortedTestResults, editor: vscode.TextEditor): void {
-    // Dots
+    // Status indicators (gutter icons)
     const styleMap = [
       {
         data: testResults.success,
-        decorationType: this.passingItStyle,
+        decorationType: this.decorations.passing,
         state: TestReconciliationState.KnownSuccess,
       },
       {
         data: testResults.fail,
-        decorationType: this.failingItStyle,
+        decorationType: this.decorations.failing,
         state: TestReconciliationState.KnownFail,
       },
       {
         data: testResults.skip,
-        decorationType: this.skipItStyle,
+        decorationType: this.decorations.skip,
         state: TestReconciliationState.KnownSkip,
       },
       {
         data: testResults.unknown,
-        decorationType: this.unknownItStyle,
+        decorationType: this.decorations.unknown,
         state: TestReconciliationState.Unknown,
       },
     ];
@@ -429,7 +423,7 @@ export class JestExt {
 
     // We have to make a new style for each unique message, this is
     // why we have to remove off of them beforehand
-    const style = decorations.failingAssertionStyle(errorMessage);
+    const style = this.decorations.failingAssertionStyle(errorMessage);
     this.failingAssertionDecorators[fileName].push(style);
 
     return { style, decorator };
@@ -502,13 +496,6 @@ export class JestExt {
       modes.push('watch');
     }
     this.status.update(status, details, modes);
-  }
-
-  private setupDecorators(): void {
-    this.passingItStyle = decorations.passingItName(this.context);
-    this.failingItStyle = decorations.failingItName(this.context);
-    this.skipItStyle = decorations.skipItName(this.context);
-    this.unknownItStyle = decorations.notRanItName(this.context);
   }
 
   private shouldIgnoreOutput(text: string): boolean {